updated structure and replaced wallclock time attribute
